Select List Fragment Done
Select List now gets all lists that the user is the owner of and all
the public lists that the user is a member of. Each list gets one line
of text with the list name. When a list is clicked, it goes to a
(currently completely wrong) select members page.
If I click on a list, go back, and then create a new list, the new list
will be named the same as the clicked list. I don’t think this will be
a problem because I plan to put the new list creation in a different
activity.