List of Locations
Finally got the list of locations. It populates the Choose Location
list with the query and location, and voila. If we click on a list
item, it saves both the venue and the name (for easy access). Then, it
goes to the Set Deadline page. The Final Edit will update, too.
Currently, it’s hardcoded to search for “Sushi Dog” around 40.7,-74. We
need to have a dynamic search and dynamic location.