In which I talk to Brian... I mean, what?
This includes the locate venue feature (though I’ve only tested it in
the browser in google maps and it works badly), a mock up of the
current invite card, and the sign in/login flow. The sign in and login
just need to correctly make the user name as a Twilio number, but they
work otherwise.