Useful Visuals
Start time and limits are now visible. Subtle view animation for when
views appear and disappear (most of the time). Add comments now looks a
little better. And you can view the invitees. And the final edit stuff
does send the right pushes (yet to be tested) and does not allow you to
edit stuff that… well, would break things.